Move runtime APIs to a separate file
